Function grpcio_sys::grpc_channel_watch_connectivity_state

pub unsafe extern "C" fn grpc_channel_watch_connectivity_state(
    channel: *mut grpc_channel, 
    last_observed_state: grpc_connectivity_state, 
    deadline: gpr_timespec, 
    cq: *mut grpc_completion_queue, 
    tag: *mut c_void
)

Watch for a change in connectivity state. Once the channel connectivity state is different from last_observed_state, tag will be enqueued on cq with success=1. If deadline expires BEFORE the state is changed, tag will be enqueued on cq with success=0.