Function grpcio_sys::grpc_channelz_get_top_channels

pub unsafe extern "C" fn grpc_channelz_get_top_channels(
    start_channel_id: isize
) -> *mut c_char

CHANNELZ API *************/ churn as the feature is implemented. This comment will be removed once channelz is officially supported, and these APIs become stable. For now you may track the progress by following this github issue: https://github.com/grpc/grpc/issues/15340

the following APIs return allocated JSON strings that match the response objects from the channelz proto, found here: https://github.com/grpc/grpc/blob/master/src/proto/grpc/channelz/channelz.proto.

For easy conversion to protobuf, The JSON is formatted according to: https://developers.google.com/protocol-buffers/docs/proto3#json.