Module grpcio_sys::grpc_status_code

Constants

GRPC_STATUS_ABORTED
The operation was aborted, typically due to a concurrency issue like sequencer check failures, transaction aborts, etc.

GRPC_STATUS_ALREADY_EXISTS
Some entity that we attempted to create (e.g., file or directory) already exists.

GRPC_STATUS_CANCELLED
The operation was cancelled (typically by the caller).

GRPC_STATUS_DATA_LOSS
Unrecoverable data loss or corruption.

GRPC_STATUS_DEADLINE_EXCEEDED
Deadline expired before operation could complete. For operations that change the state of the system, this error may be returned even if the operation has completed successfully. For example, a successful response from a server could have been delayed long enough for the deadline to expire.

GRPC_STATUS_FAILED_PRECONDITION
Operation was rejected because the system is not in a state required for the operation's execution. For example, directory to be deleted may be non-empty, an rmdir operation is applied to a non-directory, etc.

GRPC_STATUS_INTERNAL
Internal errors. Means some invariants expected by underlying system has been broken. If you see one of these errors, something is very broken.

GRPC_STATUS_INVALID_ARGUMENT
Client specified an invalid argument. Note that this differs from FAILED_PRECONDITION. INVALID_ARGUMENT indicates arguments that are problematic regardless of the state of the system (e.g., a malformed file name).

GRPC_STATUS_NOT_FOUND
Some requested entity (e.g., file or directory) was not found.

GRPC_STATUS_OK
Not an error; returned on success

GRPC_STATUS_OUT_OF_RANGE
Operation was attempted past the valid range. E.g., seeking or reading past end of file.

GRPC_STATUS_PERMISSION_DENIED
The caller does not have permission to execute the specified operation. PERMISSION_DENIED must not be used for rejections caused by exhausting some resource (use RESOURCE_EXHAUSTED instead for those errors). PERMISSION_DENIED must not be used if the caller can not be identified (use UNAUTHENTICATED instead for those errors).

GRPC_STATUS_RESOURCE_EXHAUSTED
Some resource has been exhausted, perhaps a per-user quota, or perhaps the entire file system is out of space.

GRPC_STATUS_UNAUTHENTICATED
The request does not have valid authentication credentials for the operation.

GRPC_STATUS_UNAVAILABLE
The service is currently unavailable. This is a most likely a transient condition and may be corrected by retrying with a backoff. Note that it is not always safe to retry non-idempotent operations.

GRPC_STATUS_UNIMPLEMENTED
Operation is not implemented or not supported/enabled in this service.

GRPC_STATUS_UNKNOWN
Unknown error. An example of where this error may be returned is if a Status value received from another address space belongs to an error-space that is not known in this address space. Also errors raised by APIs that do not return enough error information may be converted to this error.

GRPC_STATUS__DO_NOT_USE
Force users to include a default branch:

Type Definitions

Type