Type Definition grpcio_sys::grpc_process_auth_metadata_done_cb

type grpc_process_auth_metadata_done_cb = Option<unsafe extern "C" fn(user_data: *mut c_void, consumed_md: *const grpc_metadata, num_consumed_md: usize, response_md: *const grpc_metadata, num_response_md: usize, status: Type, error_details: *const c_char)>;

Callback function that is called when the metadata processing is done.

Consumed metadata will be removed from the set of metadata available on the call. consumed_md may be NULL if no metadata has been consumed.

Response metadata will be set on the response. response_md may be NULL.

status is GRPC_STATUS_OK for success or a specific status for an error. Common error status for auth metadata processing is either GRPC_STATUS_UNAUTHENTICATED in case of an authentication failure or GRPC_STATUS PERMISSION_DENIED in case of an authorization failure.

error_details gives details about the error. May be NULL.